Conflicts:
	src/bridge_map.h
	src/clear_map.h
	src/industry_map.h
	src/map.cpp
	src/misc_gui.cpp
	src/object_map.h
	src/rail_cmd.cpp
	src/rail_map.h
	src/road_map.h
	src/saveload/afterload.cpp
	src/saveload/map_sl.cpp
	src/saveload/oldloader_sl.cpp
	src/saveload/station_sl.cpp
	src/settings_gui.cpp
	src/station_cmd.cpp
	src/station_map.h
	src/terraform_cmd.cpp
	src/tile_map.cpp
	src/tile_map.h
	src/tile_type.h
	src/town_map.h
	src/tree_map.h
	src/tunnel_map.cpp
	src/tunnel_map.h
	src/tunnelbridge_cmd.cpp
	src/void_map.h
	src/water_map.h
	src/waypoint_cmd.cpp

+++ src/strings.cpp
@@ -1176,9 +1176,9 @@ static char *FormatString(char *buff, const char *str_arg, StringParameters *arg
 					/* Shows only passengers and mails since other cargoes provide no useful value. (all 1) */
 					if (cs->Index() == CT_PASSENGERS || cs->Index() == CT_MAIL) {
 						if (cs->Index() == CT_PASSENGERS) {
-							n = seprintf(buff, last, "(%d)", args->GetParam(2));
+							n = seprintf(buff, last, "(%d)", (int) args->GetParam(2));
 						} else {
-							n = seprintf(buff, last, "(%d)", args->GetParam(3));
+							n = seprintf(buff, last, "(%d)", (int) args->GetParam(3));
 						}
 						buff += n;
 					}
